Modify analytics/legacy/homepagevisit to account for changes on metawiki
Since we started the migration of the HomepageVisit schema to Event Platform
and created the schema in this repo, the schema in meta.wikimedia.org has
seen some changes that are not reflected in this repo. This patch updates
the schema to reflect the changes in metawiki.

The changes in meta were: 3 fields were removed. Two optional ones, and
one required one. To keep changes backwards-compatible, this change just
makes the required field that was dropped in meta -> non-required.

Bug: T267333
Change-Id: I26719ffff73200bc8ad3767f786df5f013c74f4a