Enforce numeric bounds for all schemas
This re-materializes all schemas (and includes manual edits to old versions)
with the default enforcedNumericBounds applied.

JSON does not have the concept of e.g. max Long.  Numeric values
outside of parseable ranges should be considered invalid.
All integer and number fields that do not specify maximum or minimum
will be materialized with these set to
Number.MIN_SAFE_INTEGER and Number.MAX_SAFE_INTEGER.

This is not a type change, so is sort-of backwards compatible.
It should not cause any ingestion errors.  I highly doubt it will
cause any validation errors to happen, as if this was happening
regularly Hive ingestion would be failing.

This has happened in the past to e.g. NavigationTiming
and we have manually applied a maximum there.  This just
does the same thing by default for all schemas.

Bug: T273069

Change-Id: Ibe444c0312ad8fadd41dfc32e11e13155e906d6a