universalLanguageSelector: Add new properties
Readers Web will be A/B testing the existing language switcher treatment
against a new one being worked on as part of Desktop Improvements.

Adding the session IDs, isAnon, and userEditBucket properties will allow
us to better understand baseline user behavior for both logged-in and
logged-out users, as well as logged-in power users.

Adding the selectedInterfaceLanguage field allows us to understand which
language the user selected when they switched.

Bug: T275766
Change-Id: Ifb86f62125bffa42f09243716c68d747a05fb1e8