Add new analytics/pref_diff schema
Per T261842, the web team is interested in tracking all skin changes
involving Vector including their "initial" and "final" states so that we
can see what preference the user switched from/to (the user switched
from latest Vector to Monobook). This schema differs from the prefupdate
schema in that it:

* Allows for the use of a hashed user id instead of the real user id
(for privacy reasons)

* Has an "initial_state" field that allows the "before" state of the
form to be captured.

Bug: T261842
Change-Id: I8f7058407f0205a74d6eefbd4dbbc56ed50d833e