HompageVisit: Add welcomesurvey-originalcontext to referer_route
welcomesurvey-originalcontext is used as a referer ('source' query
parameter) when the user chooses not to go to the homepage from the
welcome survey, and instead get returned to where they were before
clicking 'Create account'.

When that page is the homepage, we end up with this unexpected value
in HomepageVisit. Probably an artifact of manual testing; it's not
relevant, but add it as a recognized value anyway to keep error logs
clean.

Bug: T281232
Change-Id: Id728f48666c8d3311bcfba6a80922f22b6a9f4b3