build: Document simpler alternative contribution flow
It took me quite a while to figure what file should (not) be modified,
what command to run to generate the other files, then to undo those
changes after realizing it had overwritten the previous versions, etc.

I was expecting that jsonschema-tools would find all `current.yaml`
files, and expand them all (at least any that are new or changed).
But, it seems this is not the case. The `materialize` command is for
one-offs only, and `materialize-modified` only works for changing files
that already existed (plus needs -G to suppress a security warning).
For new files, one first needs to stage them with 'git add' and then
run the `materialize-modified` with `-s` to expand it.

It seems with the current tooling, the quickest way involving the least
steps for new files is to pass the file to `jsonschema-tools materialize`
(instead of passing it to git-add and then run another command, and
then another git-add). I've documented the two as follows:

* `npm run build-modified` - this works right away after cloning the
  repo and modifying a file. But, it does not see new files.

* `npm run build-new path/to/new/current.yaml` - for new files, which
   works right after cloning the repo and creating the subdir+file,
   without needing to one of the files first.

If the general direction is liked, we could improve jsonschema-tools
to support this with one command. For example, the default mode
of `jsonschema-tools materialize` could be to look process all
current.yaml files in the configured schemaBasePath regardless of
Git state, and expand any that differ from the latest version on disk.

Bug: T290074
Change-Id: I1082f9cc588a9709f101b605c7c1c9494e9fb60d