build: Add brief "Getting started" guide
It took me quite a while to figure what file should (not) be modified,
what command to run to generate the other files, etc. More details
about that experience in change I. I eventually realized that the
(failed) postinstall script wasn't merely running tests, but also
the thing that is meant to do all this.

I'm aware there are comprehensive pages on Wikitech covering this,
but they felt overwhelming to me a starting point for a quick fix.
They were useful as a later reference when troubleshooting, though.

The specifics of the git-precommit hook are imho something we should
work to phase out, and replace with something that managers, analysts,
and engineers can all safely use with ease. In this patch I'm merely
documenting the status quo first for first-time contributors that
are looking to modify a schema they inherited maintenance for or
when developing something anew for the first time.

Change-Id: I7fc65549ce326d6b208e12f87f69b41af696ba87