Add performer field to sparql/query
New field facilitates the user blocking functionality of WCQS, the
implementation mimics how users are represented in the
mediawiki/common fragment but limited to the name as it's all we have
available.

Also corrects the most recent changelog entry (from I51c07442) which
released version 1.1.0 but left a changelog entry for 1.0.1.

Bug: T293462
Change-Id: I36ca66d8b1a289a5bc1a853e49bf36e8cd2ea7a1