Mediasearch_Interaction: Add search_result_page_title field
Previously, whenever a user triggered the Quickview feature in
MediaSearch, a "result_click" event would be logged, and a
"search_result_page_id" field was included in the data payload.
The implementation of the Quickview feature has changed, and MW
pageIDs are no longer used as identifiers internally.

This change updates the schema to add a new field,
search_result_page_title, that can be used in place of the
search_result_page_id field.

Bug: T297400
Change-Id: I329337b63c133974a0a3337f061df3d684c1e95d