mediawiki/client/metrics_event: Make custom_data a map type
In I3afe7a5e493, the custom_data field was defined as a schemaless
object, which is not supported [0]. Update the schema for the field to
be a map type.

Note well that the mediawiki/client/metrics_event schema ID has not been
bumped despite changing the type of a field. This is acceptable in this
instance because there are no downstream clients of the few events
logged using the schema, which will be dropped once the associated
instrument has been disabled (see I6fdfb563a0a).

[0] https://wikitech.wikimedia.org/wiki/Event_Platform/Schemas/Guidelines#map_types

Bug: T314151
Depends-On: I6fdfb563a0a9609c74ed79db255ee6138c63c117
Change-Id: I399912449c4811209bdaa9fd59a61ef177cc1d95