content_interaction_event: Fix typo and add new val to enum
This retroactively fixes a typo that should not have been present.
(Note: the corresponding instrumentation does NOT have the typo,
 so those events would get rejected by EventGate as invalid. This
 fixes that.)

This also adds a new possible value to the same enum -- 'N/A' which
would be the case for logged-out users. Right now those events have
it as `null` and the corresponding instrumentation would need to be
updated to use the new enum value ('N/A').

Bug: T299055
Change-Id: I2d3b52c704f88ca9818677aae13ff57a0a14c273