Add analytics/mediawiki/mentor_dashboard/personalized_praise
Originally, I thought about creating a more generic module,
similar to homepagemodule legacy module. Eventually, I decided
against it, because the majority of events regarding personalized_praise
are not user interaction with the module, but system actions,
so it makes little sense to log those two types of actions
together.

If we want to eventually know which links mentor tend to click
on the dashboard, that'd be a great job for a potential future
analytics/mediawiki/mentor_dashboard/module schema.

This schema will be used in GrowthExperiments as part of
in-development feature called Personalized praise.

Bug: T325117
Change-Id: I88972036c690bdbb4845283a75048016375b473d