Add new unified mobile apps schema for Session
This PR adds a new unified schema to track sessions in both iOS and Android apps.
It's based on Android's existong schema android_app_session
It adds a field for the iOS required fiel primary_language, while makes the languages fields non-requires, as the iOS app does not send it
Also, it adds a comment on the existing schemas warning of their deprecation.
It adds a missing field to the iOS search and reading lists schemas and a missing action to the iOS login schema

Bug: T331481
Change-Id: I8e684abd58310299fd7eaa1395b1cb033de4bf9f