Update schemas for iOS diff view changes
This adds the ios_watchlists schema, which so far only includes changes needed for the iOS diff view.
Additional watchlist-specific actions will be added at a later time.
This also adds a category to ios_login_action, so that we can track login and create account events
triggered from the diff view.

Bug: T341896
Change-Id: I73adf2311c6aab2232f712a41189d2c51612f8a8