Add useragent to eventlogging legacy fragement eventcapsule
In order to populate the useragent using the new version of the Refine
process. We have the option of declaring the useragent struct
directly into the schema. This will allow us to simplify the
code by ignoring this legacy case.

This first patch adds the useragent legacy struct into a new
version of the eventcapsule fragment. The next patch will bump
the used version of the fragment into the actual schemas.

Bug: T356762
Change-Id: Ib2bbb6ea659733e486faadff2bf36b147bd2aa3a