Add useragent to analytics/legacy events
In order to populate the useragent using the new version of the Refine
process. We have the option of declaring the useragent struct
directly into the schema. This will allow us to simplify the
code by ignoring this legacy case.

The first patch 1075636 added the useragent legacy struct into a new
version of the eventcapsule fragment. This patch bumps
the used version of the fragment into the actual schemas.

Bug: T356762
Change-Id: I4860b3c521b986b67a5d40f1fdaaa65324a1b71d